Revert "run-dev: Run process_queue with DJANGO_AUTORELOAD_ENV."
This reverts commit 3acb91cd7a947864f8165bdcc0421e6bf58e5530 (#13934).

The Django 2.2 autoreloader works by forking into a child process that
exits with status 3 when a file changes, and a parent process that
restarts the child when it exits with status 3.  Setting this
environment variable had the effect of pretending we were already the
child process, without a parent process to restart it.  Therefore,
changing any code used by the queue processor caused it to exit rather
than restart.

Signed-off-by: Anders Kaseorg <anders@zulipchat.com>