message_list: Use document.getSelection to identify drags.
This removes the messy click-or-drag detection algorithm originally
added in a7b7ea2dfb56f610f4bc2b068e33da59a20b8d3f, which fixed a messy
bug in an earlier algorithm from ~2013, whose sole purpose as to check
whether we're doing a selection and if so, not trigger the
click-on-message-body click handler.

The right fix is of course to do that check correctly.