stream_topic_history: Fix incorrectly migrated field name.
In 97ead5f397f01c8d9de901bc939850bf889d6bdb, we renamed the topic_date
-> stream_topic_history, and in the process renamed some local object
properties from .name => .topic_name, and accidentally change the
type for the data from the server as well.

The test fixtures were incorrectly migrated in the same way, so we fix
that as well.