realm logo: Display logo full width regardless of the size of the image.
Display logo at full width regardless of the size of the image to
reduce the dependency on the logo image in determining the logo
container size.  This also fixes a problem owhere the night/day logos
would lose their default-dark/white background color when we upload an
image in jpg format rather than png.