search_box: Fix search bar background on nightmode.
Commit eea943a50e94fbebbae0b1fb8d8ffe896c5a3d00 introduced a
regression that caused a small part of the navbar in night mode to not
have the correct background color.

The relevant changes in that commit intended to fix the margin for the
search box for when the search pills feature was set to active.

This commit slightly increases the padding for the search box (when
pills are active), to improve pill alignment, and adds styles for
"#searchbox_legacy" to correct the background when search pills are
disabled.

This also reverts the change from commit
ce1df7e7938c8bbcd322ec441ffcf79f5b2d330d which tried to improve the
alignment of pills by adding a margin left but didn't address the
background color issue.